Switch to using injectable to setup route for getting service account
