+++ src/main/routes/index.ts
@@ -3,5 +3,4 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 
-export * from "./kubeconfig-route";
 export * from "./version-route";
